Fixed speed bug
Previously a fixed-size array was used to calculate the max speed for ghosts, meaning that if you beat too many levels there would be errors. Now the random speed is just calculated to be 1-4 for random values 0-3, and multiples of 2 for larger values:
rand of 4=6 speed, 5->8, 6->10 (extrapolated from the array)

Additionally, running the game from the wrong directory would cause the sprites not to load, so they now use a relative path to the resources folder. I also made multiple stylistic changes to match my preferences.